requirements.txt
langchain>=0.1.0
neo4j>=5.14.0
transformers>=4.36.0
torch>=2.1.0
sentence-transformers>=2.2.2
python-dotenv>=1.0.0
llama-cpp-python>=0.2.23
chromadb>=0.4.22
pyyaml>=6.0.1
pytest>=8.3.4
black>=23.12.1
isort>=5.13.2
psutil>=6.1.1